docshell/test/mochitest/test_close_onpagehide_by_window_close.html

<!doctype html>
<title>Test for closing window in pagehide event callback caused by window.close()</title>
<script src="/tests/SimpleTest/SimpleTest.js"></script>
<link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=1432396">Mozilla Bug 1432396</a>
<p id="display"></p>
<script>
SimpleTest.waitForExplicitFinish();

const w = window.open("file_close_onpagehide1.html");
window.addEventListener("message", e => {
  is(e.data, "initial", "The initial page loaded");
  w.onpagehide = () => {
    w.close();
    info("try to close the popped up window in onpagehide");
    SimpleTest.finish();
  };
  w.close();
}, { once: true });
</script>
